modified kali virtual guestadditions script into two scripts to support vbox and vmware. added confirmation dialogue to reboots

+++ kali-vbox.sh
@@ -17,9 +17,19 @@ if [ ! -f $step_one ]; then
 	echo '[*] installing guest additions...'
 	apt-get install -y virtualbox-guest-x11
 
-	echo '[*] rebooting to complete install...'
-	touch $step_one
-	reboot
+	# confirmation thanks to https://stackoverflow.com/a/1885534
+	read -p "Would you like to reboot to continue install? " -n 1 -r
+	echo    # (optional) move to a new line
+	if [[ $REPLY =~ ^[Yy]$ ]]
+	then
+	    echo '[*] rebooting to complete install...'
+		touch $step_one
+		reboot
+
+	fi
+	# we didn't confirm
+	echo '[!] reboot aborted!'
+	exit 1
 else
 	echo '[+] step one already complete. moving to step two'
 fi
@@ -28,9 +38,19 @@ if [ ! -f $step_two ]; then
 	echo '[*] upgrading the dist'
 	apt-get dist-upgrade
 
-	echo '[*] rebooting to complete install...'
-	touch $step_two
-	reboot
+	# confirmation thanks to https://stackoverflow.com/a/1885534
+	read -p "Would you like to reboot to continue install? " -n 1 -r
+	echo    # (optional) move to a new line
+	if [[ $REPLY =~ ^[Yy]$ ]]
+	then
+	    echo '[*] rebooting to complete install...'
+		touch $step_two
+		reboot
+
+	fi
+	# we didn't confirm
+	echo '[!] reboot aborted!'
+	exit 1
 else
 	echo '[+] step two already complete'
 fi
